Removes code duplication in the databases
Fixes a regression bug on database connections - more than 1 were being
made again.
