+++ src/main/java/world/bentobox/bentobox/database/sql/mariadb/MariaDBDatabaseHandler.java
@@ -2,7 +2,7 @@ package world.bentobox.bentobox.database.sql.mariadb;
 
 import world.bentobox.bentobox.BentoBox;
 import world.bentobox.bentobox.database.DatabaseConnector;
-import world.bentobox.bentobox.database.sql.AbstractSQLDatabaseHandler;
+import world.bentobox.bentobox.database.sql.SQLDatabaseHandler;
 import world.bentobox.bentobox.database.sql.SQLConfiguration;
 
 /**
@@ -13,7 +13,7 @@ import world.bentobox.bentobox.database.sql.SQLConfiguration;
  *
  * @param <T>
  */
-public class MariaDBDatabaseHandler<T> extends AbstractSQLDatabaseHandler<T> {
+public class MariaDBDatabaseHandler<T> extends SQLDatabaseHandler<T> {
 
     /**
      * Handles the connection to the database and creation of the initial database schema (tables) for
