+++ src/main/java/world/bentobox/bentobox/database/sql/postgresql/PostgreSQLDatabaseHandler.java
@@ -2,7 +2,7 @@ package world.bentobox.bentobox.database.sql.postgresql;
 
 import world.bentobox.bentobox.BentoBox;
 import world.bentobox.bentobox.database.DatabaseConnector;
-import world.bentobox.bentobox.database.sql.AbstractSQLDatabaseHandler;
+import world.bentobox.bentobox.database.sql.SQLDatabaseHandler;
 import world.bentobox.bentobox.database.sql.SQLConfiguration;
 
 /**
@@ -12,7 +12,7 @@ import world.bentobox.bentobox.database.sql.SQLConfiguration;
  * @since 1.6.0
  * @author tastybento, Poslovitch
  */
-public class PostgreSQLDatabaseHandler<T> extends AbstractSQLDatabaseHandler<T> {
+public class PostgreSQLDatabaseHandler<T> extends SQLDatabaseHandler<T> {
 
     /**
      * Constructor
@@ -22,7 +22,7 @@ public class PostgreSQLDatabaseHandler<T> extends AbstractSQLDatabaseHandler<T> 
      *                          values from the database or inserted into the database
      * @param databaseConnector Contains the settings to create a connection to the database
      */
-    protected PostgreSQLDatabaseHandler(BentoBox plugin, Class<T> type, DatabaseConnector databaseConnector) {
+    PostgreSQLDatabaseHandler(BentoBox plugin, Class<T> type, DatabaseConnector databaseConnector) {
         super(plugin, type, databaseConnector, new SQLConfiguration(type.getCanonicalName()));
     }
